Added some functions, the simulation is now graphically represented
The functions added are the exponential_distribution and the functions
that initialize and update the time state to follow some given
distribution: as of now, that distribution is the exponential, and that's
why it was implemented in the first place.
Moreover, now run_simulation() represents the results of the simulation
by means of matplotlib, with two graphs where are represented the evolution
of x and y as functions of t and the trajectory (y and x points).
All of these modifications where minimally tested