Big update introducing proper CTRW and Bacterium class
The logic of the previous functions has been changed to operate on
single values instead of lists (taking advantage of the simulation
being of the Markov chain's kind so memorylessness is a feature and
the simulation doesn't need to know the whole past of the system to
correctly operate).
Introduced a run_simulation() function that properly simulates a
continuous time random walk for an instance of the new Bacterium
class, which may move, die or reproduce.
As of now, movement is implemented as an update of the space
coordinates, death is just a simulation loop break and reproduction
does virtually nothing (the code to simulate more bacteria will come
later, for now reproduction is a stub)

Also changed the distributions and functions dependent on random
number's generation since a bug was spotted in their previous form
where pseudonumbers weren't properly updated, now those functions
are called passing the default generator of numpy and (says my
test suite) the functions should be reproducible

The code is yet minimally tested by means of basic unit testing

Also changed in the main code and in tests.py the format of the
docstrings, that still must be updated since most are empty