Big update separating analysis and simulation
brownie_box.py has been separated into two different files, simulation.py
and analysis.py. Now the first script is run and produces the output of the
simulation in a json file containing the simulation parameters and the
bacterium states.
Then files of this kind are fed into analysis.py which loads the data,
parses it and then plots it.

The data are stored as dictionaries since it is the most natural form to
use with a json hierarchical file, so in the simulation.py code the
Bacterium class has been suppressed in favor of dictionaries to store
the bacterium states and data.

Tests have also been updated to account for this update and as usual for
now the code is tested very minimally. Readme file has also been
updated with the new usage info