Added multiple bacteria and reproduction dynamics
Now the simulation can handle multiple bacteria at once, which are initialized by the
user at the start of the simulation. A new kind of event, 'birth', has been added
to reflect the first initialization of the bacterium. Reproduction has been
implemented, now when the event reproduction is drawn, the bacterium will stop in its
place and 'spawn' another bacterium in the same position it occupies, which will have
the event 'birth' in it and will be subject to simulation the next turn of the loop.

tests.py and analysis.py have been modified to reflect this upgrade, especially the
latter where now the plot of trajectory and space-time evolution must be selected
by filename of the simulation and bacterium index in the list of bacteria loaded
from the json file