Added death, reproduction and movement parameters
When the simulation is called these three parameters are now added to the coefficients
of the simulation, both in the dictionary of the parameters of simulation and in
the function call... the functions for death, reproduction and movement now take
a parameter as input, so that in the future if their functional form has to
become more complex, it should be possible to upgrade them without problem. And even
if their functional form becomes more complex, it makes sense that they should be
proportional to a constant that alters the probability of that process and that said
constant should be registered as the other parameters of simulation.