Big update - bacteria list management and checks
The code underwent big updates for logical soundness, stability and
optimization of resource usage

The first main concern was to split the active, dead and inactive
bacteria into separate lists, so that the loops in the main simulation
avoid doing manipulation to dead and inactive bacteria. In this way once
bacteria die or stop being evolved due to time limit, they are moved in lists
that store them for later saving them, meanwhile if clauses and calculations
are performed only on the active bacteria, saving up a lot of CPU usage
and speeding up the performance

The previous code made use of counters to track dead or active bacteria from
the single list they were stored. Since now there are lists for every each
case those counters became useless and the checks and manipulations are now
done by means of lists, which is also a better suited approach to an
object oriented language such as python; as an example, to check if all
bacteria died in the previous iterations of this script a counter was used
and compared to the number of total bacteria. Now the same check is performed
by checking if the active and inactive lists are both empty, meaning that
the whole bacteria is in the dead list

A clearer output of the simulation running was added, which updates in real
time by giving the user the number of alive bacteria over the bacteria limit
and the time reached by the simulation over the time limit; this becomes
useful to check at runtime the software behavior or any possible problems

The bacteria max threshold was changed from a soft limit to a hard limit. In
previous versions this value was checked only at the end of the turn, aka
the end of a while loop iteration, and it was intended only as a safety
measure to avoid diverging simulation times that weren't catched by other
checks and that were due to the exponential reproduction mechanism of
the bacteria. Since it was checked at the end of the turn during that turn
the limit number could be surpassed without the simulation stopping, since
this happened during the for loop iterations that evolve every single
bacteria that all happen in a single turn (while loop iteration). This
was intentional, since checking this threshold inside the for loop would
have meant having an if clause check for a lot of bacteria every turn
and that was too expensive computationally.
But now since inactive and dead bacteria are excluded from these for loops
the checks are far less expensive and were moved in the for loop, making
the bacteria limit a hard threshold that once reached exactly will be
enforced. This choice felt logically in line with the other limit, the time
limit, which was already a hard threshold

Various comment were added or edited to reflect these changes and tell
expicitly some of this info to the user; some main names and other minor
edits were done